Feat: 사이드바에 PC용 버튼 추가하여 URL 파라미터에 단계 표시되도록 설정
