+++ src/app/addform/page.tsx
@@ -1,16 +1,10 @@
 import StepSidebar from "./components/StepSidebar";
 
-interface AddFormPageProps {
-  searchParams: Promise<{ step: string }>;
-}
-
-const AddFormPage = async ({ searchParams }: AddFormPageProps) => {
-  const { step } = await searchParams;
-
+const AddFormPage = async () => {
   return (
-    <div>
-      <StepSidebar step={step} />
-    </div>
+    <>
+      <StepSidebar />
+    </>
   );
 };
